+++ nebula/client_test.go
@@ -1,10 +1,16 @@
-package ngdb
+/* Copyright (c) 2019 vesoft inc. All rights reserved.
+ *
+ * This source code is licensed under Apache 2.0 License,
+ * attached with Common Clause Condition 1.0, found in the LICENSES directory.
+ */
+
+package nebula
 
 import (
 	"fmt"
 	"testing"
 
-	"github.com/vesoft-inc/nebula-go/graph"
+	"github.com/vesoft-inc/nebula-go/nebula/graph"
 )
 
 const (
